chore: test improvements
