+++ test/fixtures/hooks/error-hooks/terraform.sh
@@ -1,0 +1,3 @@
+#!/bin/sh
+
+(set -x && exec "terraform" "$@" 2>&1)
